moving measure enrichment to cover hard-code measures.
